Move root logger to module level
