+++ new_ffrd_event_item.py
@@ -1,6 +1,6 @@
 from pystac import Item
 
-from hecstac.common.logger import initialize_logger
+from hecstac.events.logger import initialize_logger
 from hecstac.events.ffrd import EventItem
 
 if __name__ == "__main__":
